dia_01/onboard/requirements.txt
ibm-watsonx-ai>=1.1.0
elasticsearch==8.11.1
pymilvus>=2.4.0
jupyterlab>=4.2.0
numpy>=1.26.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
